Refactor: Modularize LSP server handlers (formatted)
