--- src/lsp/handler.rs
@@ -1,1 +1,0 @@
-// LSP request handlers for Hover, Go to Definition, etc.
